refactor query under achievement context
